# opspilot/infra/openapi.json
{"openapi":"3.1.0","info":{"title":"OpsPilot API","description":"Multi-Agent Autonomous SRE Command Center for Enterprise Cloud Operations. Powered by LangGraph, Azure AI Foundry, and GPT-4o.","version":"0.2.0"},"paths":{"/api/incidents/active":{"get":{"tags":["incidents"],"summary":"List active incidents","description":"Returns all incidents in OPEN or INVESTIGATING status.","operationId":"get_active_incidents_api_incidents_active_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/IncidentRecord"},"type":"array","title":"Response Get Active Incidents Api Incidents Active Get"}}}}}}},"/api/incidents/{incident_id}":{"get":{"tags":["incidents"],"summary":"Get incident by ID","description":"Retrieves a single incident record by its unique identifier.","operationId":"get_incident_api_incidents__incident_id__get","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentRecord"}}}},"404":{"description":"Incident not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/incidents/":{"post":{"tags":["incidents"],"summary":"Create incident and trigger investigation","description":"Creates a new incident record and immediately starts a background investigation via the InvestigationOrchestrator. The incident is available via GET /api/incidents/{id} and its SSE stream via GET /api/incidents/{id}/stream as soon as this endpoint returns.","operationId":"create_incident_api_incidents__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateIncidentRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentRecord"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/recommendations/{incident_id}":{"get":{"tags":["recommendations"],"summary":"Get recommendations for an incident","description":"Returns the confirmed root cause analysis and prioritised remediation actions generated by the Commander agent for the given incident.","operationId":"get_recommendations_api_recommendations__incident_id__get","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationResponse"}}}},"404":{"description":"Incident not found or analysis not yet complete"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/incidents/{incident_id}/stream":{"get":{"tags":["stream"],"summary":"Stream agent activity events (SSE, read-only)","description":"Server-Sent Events stream of real-time agent activity for an incident. Read-only: it streams an already-running investigation and never starts one. Use POST /api/incidents/{id}/investigate to start a run.","operationId":"stream_incident_api_incidents__incident_id__stream_get","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}},"text/event-stream":{}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/incidents/{incident_id}/investigate":{"post":{"tags":["investigations"],"summary":"Re-run the full agent investigation","description":"Launches the InvestigationOrchestrator (commander → metrics → logs → deployment → time_machine → root_cause → [deep_reasoning] → recommendation) as a background task. Real agent execution; progress is emitted over the incident's SSE stream. No-op if a run is already in flight.","operationId":"investigate_api_incidents__incident_id__investigate_post","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvestigateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/incidents/{incident_id}/deep-reasoning":{"post":{"tags":["investigations"],"summary":"Run deep reasoning (o4-mini) over the current findings","description":"Sends the specialist findings + current root cause to the DeepReasoningAgent (REASONING/o4-mini role) and returns a refined root cause with a chain-of-thought reasoning trace. Real model execution in foundry mode; emits SSE progress events.","operationId":"deep_reasoning_api_incidents__incident_id__deep_reasoning_post","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeepReasoningRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeepReasoningResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/investigations":{"get":{"tags":["insights"],"summary":"Completed investigation history","operationId":"list_investigations_api_investigations_get","parameters":[{"name":"incident_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InvestigationRecord"},"title":"Response List Investigations Api Investigations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/investigations/latest":{"get":{"tags":["insights"],"summary":"Most recent investigation record","operationId":"latest_investigation_api_investigations_latest_get","parameters":[{"name":"incident_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/InvestigationRecord"},{"type":"null"}],"title":"Response Latest Investigation Api Investigations Latest Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics":{"get":{"tags":["insights"],"summary":"Analytics computed from stored investigations","operationId":"analytics_api_analytics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Analytics Api Analytics Get"}}}}}}},"/api/agents/stats":{"get":{"tags":["insights"],"summary":"Per-agent execution stats from stored investigations","operationId":"agent_stats_api_agents_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response Agent Stats Api Agents Stats Get"}}}}}}},"/api/system/health":{"get":{"tags":["system"],"summary":"Azure AI Foundry configuration health","description":"Returns current Foundry configuration status, model deployment names, and the list of available investigation agents. executionMode is 'live' when AZURE_OPENAI_ENDPOINT is set, 'mock' otherwise.","operationId":"system_health_api_system_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FoundryHealthResponse"}}}}}}},"/api/agents/test":{"post":{"tags":["system"],"summary":"Run specialist agents against a test incident","description":"Runs MetricsAgent, LogsAgent, and DeploymentAgent concurrently against the supplied incident description. When AZURE_OPENAI_ENDPOINT is set the agents call Azure OpenAI; otherwise they return deterministic mock findings. All three agents are always executed regardless of Foundry configuration.","operationId":"test_agents_api_agents_test_post","parameters":[{"name":"X-API-KEY","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentTestRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentTestResponse"}}}},"401":{"description":"Invalid or missing X-API-KEY (when DEV_API_KEY is set)"},"422":{"description":"Incident description too short (min 10 characters)"}}}},"/api/system/services":{"get":{"tags":["system"],"summary":"Monitored services health roster","description":"Returns the health of every service OpsPilot monitors — service name, health status, last incident timestamp, and representative response time. Data comes from the active TelemetryProvider: deterministic fixtures when TELEMETRY_MODE=synthetic, or real Application Insights / Log Analytics telemetry when TELEMETRY_MODE=azure.","operationId":"list_monitored_services_api_system_services_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MonitoredServicesResponse"}}}}}}},"/api/test/foundry":{"post":{"tags":["test"],"summary":"Smoke-test the AI provider layer","description":"Routes the supplied message through the active provider (selected by EXECUTION_MODE) using the requested model role. Use this to verify the provider layer independently of the agents.","operationId":"test_foundry_api_test_foundry_post","parameters":[{"name":"X-API-KEY","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FoundryTestRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FoundryTestResponse"}}}},"401":{"description":"Invalid or missing X-API-KEY (when DEV_API_KEY is set)"},"503":{"description":"Foundry mode selected but not configured"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}}},"components":{"schemas":{"AgentExecution":{"properties":{"role":{"type":"string","title":"Role"},"role_label":{"type":"string","title":"Role Label"},"status":{"type":"string","title":"Status","default":"complete"},"confidence":{"type":"number","title":"Confidence","default":0.0},"duration_seconds":{"type":"number","title":"Duration Seconds","default":0.0},"finding":{"type":"string","title":"Finding","default":""},"evidence":{"items":{"type":"string"},"type":"array","title":"Evidence"},"started_at":{"type":"string","title":"Started At","default":""},"completed_at":{"type":"string","title":"Completed At","default":""}},"type":"object","required":["role","role_label"],"title":"AgentExecution"},"AgentRunResult":{"properties":{"agent":{"type":"string","title":"Agent"},"roleLabel":{"type":"string","title":"Rolelabel"},"mode":{"type":"string","title":"Mode","description":"'live' or 'mock'"},"confidence":{"type":"number","title":"Confidence"},"summary":{"type":"string","title":"Summary"},"evidence":{"items":{"type":"string"},"type":"array","title":"Evidence"},"durationMs":{"type":"number","title":"Durationms"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata"}},"type":"object","required":["agent","roleLabel","mode","confidence","summary","evidence","durationMs","metadata"],"title":"AgentRunResult"},"AgentTestRequest":{"properties":{"incident":{"type":"string","minLength":10,"title":"Incident","description":"Plain-text incident description to investigate","examples":["Checkout service failing after deployment"]}},"type":"object","required":["incident"],"title":"AgentTestRequest"},"AgentTestResponse":{"properties":{"incident":{"type":"string","title":"Incident"},"timestamp":{"type":"string","title":"Timestamp"},"foundryConfigured":{"type":"boolean","title":"Foundryconfigured"},"results":{"items":{"$ref":"#/components/schemas/AgentRunResult"},"type":"array","title":"Results"},"totalDurationMs":{"type":"number","title":"Totaldurationms"}},"type":"object","required":["incident","timestamp","foundryConfigured","results","totalDurationMs"],"title":"AgentTestResponse"},"CreateIncidentRequest":{"properties":{"description":{"type":"string","maxLength":2000,"minLength":10,"title":"Description"},"affected_services":{"items":{"type":"string"},"type":"array","title":"Affected Services"},"reported_severity":{"$ref":"#/components/schemas/IncidentSeverity","default":"P1"},"reporter":{"type":"string","title":"Reporter","default":"anonymous"}},"type":"object","required":["description"],"title":"CreateIncidentRequest"},"DeepReasoningRequest":{"properties":{"incident_description":{"type":"string","title":"Incident Description","description":"Incident description","default":""},"findings":{"items":{"$ref":"#/components/schemas/FindingInput"},"type":"array","title":"Findings","description":"Specialist findings (metrics/logs/deployment) to reason over"},"root_cause":{"anyOf":[{"$ref":"#/components/schemas/FindingInput"},{"type":"null"}],"description":"Current (low-confidence) root cause to refine"}},"type":"object","title":"DeepReasoningRequest"},"DeepReasoningResponse":{"properties":{"incident_id":{"type":"string","title":"Incident Id"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"confidence":{"type":"number","title":"Confidence"},"blast_radius":{"type":"integer","title":"Blast Radius"},"affected_users":{"type":"integer","title":"Affected Users"},"hourly_impact_usd":{"type":"number","title":"Hourly Impact Usd"},"evidence":{"items":{"type":"string"},"type":"array","title":"Evidence"},"reasoning_trace":{"type":"string","title":"Reasoning Trace"},"mode":{"type":"string","title":"Mode"}},"type":"object","required":["incident_id","title","description","confidence","blast_radius","affected_users","hourly_impact_usd","evidence","reasoning_trace","mode"],"title":"DeepReasoningResponse"},"FindingInput":{"properties":{"role":{"type":"string","title":"Role","description":"metrics | logs | deployment | root_cause | …"},"summary":{"type":"string","title":"Summary","default":""},"evidence":{"items":{"type":"string"},"type":"array","title":"Evidence"},"confidence":{"type":"number","title":"Confidence","default":0.0}},"type":"object","required":["role"],"title":"FindingInput"},"FoundryHealthResponse":{"properties":{"foundryConfigured":{"type":"boolean","title":"Foundryconfigured"},"specialistModel":{"type":"string","title":"Specialistmodel"},"commanderModel":{"type":"string","title":"Commandermodel"},"reasoningModel":{"type":"string","title":"Reasoningmodel"},"agentsAvailable":{"items":{"type":"string"},"type":"array","title":"Agentsavailable"},"executionMode":{"type":"string","title":"Executionmode","description":"'live' when Azure OpenAI is configured, 'mock' otherwise"},"azureOpenAiEndpoint":{"type":"string","title":"Azureopenaiendpoint","description":"Endpoint URL (masked) or empty string if not set"}},"type":"object","required":["foundryConfigured","specialistModel","commanderModel","reasoningModel","agentsAvailable","executionMode","azureOpenAiEndpoint"],"title":"FoundryHealthResponse"},"FoundryTestRequest":{"properties":{"message":{"type":"string","minLength":1,"title":"Message","description":"Prompt to send through the provider.","examples":["Analyze checkout-service outage"]},"role":{"type":"string","title":"Role","description":"Model role to route to: commander | specialist | reasoning.","default":"commander"}},"type":"object","required":["message"],"title":"FoundryTestRequest"},"FoundryTestResponse":{"properties":{"provider":{"type":"string","title":"Provider","description":"'mock' or 'foundry'"},"model":{"type":"string","title":"Model","description":"Deployment name that backed the role"},"role":{"type":"string","title":"Role"},"response":{"type":"string","title":"Response"}},"type":"object","required":["provider","model","role","response"],"title":"FoundryTestResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthStatus":{"type":"string","enum":["healthy","degraded","unhealthy","unknown"],"title":"HealthStatus","description":"Coarse service health, derived from golden-signal telemetry."},"ImpactLevel":{"type":"string","enum":["critical","high","medium","low"],"title":"ImpactLevel"},"IncidentRecord":{"properties":{"id":{"type":"string","title":"Id"},"description":{"type":"string","title":"Description"},"status":{"$ref":"#/components/schemas/IncidentStatus","default":"open"},"severity":{"$ref":"#/components/schemas/IncidentSeverity"},"affected_services":{"items":{"type":"string"},"type":"array","title":"Affected Services"},"reporter":{"type":"string","title":"Reporter"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"resolved_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Resolved At"},"langgraph_run_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Langgraph Run Id"},"error_rate_pct":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Error Rate Pct"}},"type":"object","required":["id","description","severity","affected_services","reporter","created_at","updated_at"],"title":"IncidentRecord"},"IncidentSeverity":{"type":"string","enum":["P0","P1","P2","P3"],"title":"IncidentSeverity"},"IncidentStatus":{"type":"string","enum":["open","investigating","mitigated","resolved","post_mortem"],"title":"IncidentStatus"},"InvestigateResponse":{"properties":{"incident_id":{"type":"string","title":"Incident Id"},"status":{"type":"string","title":"Status","description":"'started' | 'already_running'"},"mode":{"type":"string","title":"Mode","description":"'live' (real o4-mini) | 'mock'"}},"type":"object","required":["incident_id","status","mode"],"title":"InvestigateResponse"},"InvestigationRecord":{"properties":{"id":{"type":"string","title":"Id"},"incident_id":{"type":"string","title":"Incident Id"},"description":{"type":"string","title":"Description","default":""},"started_at":{"type":"string","title":"Started At","default":""},"completed_at":{"type":"string","title":"Completed At","default":""},"duration_seconds":{"type":"number","title":"Duration Seconds","default":0.0},"status":{"type":"string","title":"Status","default":"complete"},"mode":{"type":"string","title":"Mode","default":"mock"},"severity":{"type":"string","title":"Severity","default":""},"combined_confidence":{"type":"number","title":"Combined Confidence","default":0.0},"escalated":{"type":"boolean","title":"Escalated","default":false},"root_cause":{"additionalProperties":true,"type":"object","title":"Root Cause"},"recommendations":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Recommendations"},"agents":{"items":{"$ref":"#/components/schemas/AgentExecution"},"type":"array","title":"Agents"}},"type":"object","required":["id","incident_id"],"title":"InvestigationRecord"},"MonitoredServicesResponse":{"properties":{"telemetryMode":{"type":"string","title":"Telemetrymode","description":"Active telemetry source ('synthetic' | 'azure')"},"services":{"items":{"$ref":"#/components/schemas/ServiceHealth"},"type":"array","title":"Services"}},"type":"object","required":["telemetryMode","services"],"title":"MonitoredServicesResponse"},"RecommendationResponse":{"properties":{"incident_id":{"type":"string","title":"Incident Id"},"root_cause":{"$ref":"#/components/schemas/RootCause"},"actions":{"items":{"$ref":"#/components/schemas/RecommendedAction"},"type":"array","title":"Actions"}},"type":"object","required":["incident_id","root_cause","actions"],"title":"RecommendationResponse","description":"Full recommendation payload for an incident."},"RecommendedAction":{"properties":{"id":{"type":"string","title":"Id"},"incident_id":{"type":"string","title":"Incident Id"},"priority":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Priority"},"type":{"type":"string","title":"Type"},"type_label":{"type":"string","title":"Type Label"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"steps":{"items":{"type":"string"},"type":"array","title":"Steps"},"risk":{"$ref":"#/components/schemas/RiskLevel"},"risk_label":{"type":"string","title":"Risk Label"},"impact":{"$ref":"#/components/schemas/ImpactLevel"},"impact_label":{"type":"string","title":"Impact Label"},"estimated_time":{"type":"string","title":"Estimated Time"}},"type":"object","required":["id","incident_id","priority","type","type_label","title","description","risk","risk_label","impact","impact_label","estimated_time"],"title":"RecommendedAction","description":"A single prioritised remediation action."},"RiskLevel":{"type":"string","enum":["safe","medium","high","critical"],"title":"RiskLevel"},"RootCause":{"properties":{"incident_id":{"type":"string","title":"Incident Id"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"confidence":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Confidence"},"blast_radius":{"type":"integer","minimum":0.0,"title":"Blast Radius","description":"Number of services directly affected"},"affected_users":{"type":"integer","minimum":0.0,"title":"Affected Users"},"hourly_impact_usd":{"type":"number","minimum":0.0,"title":"Hourly Impact Usd"},"evidence":{"items":{"type":"string"},"type":"array","title":"Evidence"}},"type":"object","required":["incident_id","title","description","confidence","blast_radius","affected_users","hourly_impact_usd"],"title":"RootCause","description":"Confirmed root cause identified by the Commander agent."},"ServiceHealth":{"properties":{"name":{"type":"string","title":"Name","description":"Service / Container App name"},"status":{"$ref":"#/components/schemas/HealthStatus","description":"Coarse health classification"},"responseTimeMs":{"type":"number","title":"Responsetimems","description":"Representative response time (p99 latency, milliseconds)"},"errorRatePct":{"type":"number","title":"Errorratepct","description":"Current error rate as a percentage","default":0.0},"lastIncident":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lastincident","description":"ISO-8601 timestamp of the most recent incident, or null if none"},"source":{"type":"string","title":"Source","description":"Telemetry source that produced this row ('synthetic' | 'azure')"}},"type":"object","required":["name","status","responseTimeMs","source"],"title":"ServiceHealth","description":"One row in the dashboard's 'Monitored Services' panel."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}